Crate mac_notification_sys

A very thin wrapper around NSNotifications

Modules

error

Custom errors for mac-notification-sys.

Structs

Notification

Options to further customize the notification

Enums

MainButton

Possible actions accessible through the main button of the notification

NotificationResponse

Response from the Notification

Sound

Helper to determine whether you want to play the default sound or custom one

Functions

get_bundle_identifier

Search for a BundleIdentifier of an given appname.

get_bundle_identifier_or_default

Search for a possible BundleIdentifier of a given appname. Defaults to “com.apple.Finder” if no BundleIdentifier is found.

send_notification

Delivers a new notification

set_application

Set the application which delivers or schedules a notification